% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{Features}
\alias{Features}
\title{Features Set}
\arguments{
\item{features_config}{AWS S3 Tiledb config}

\item{data}{X field of Ohlcv object}

\item{x}{List}

\item{FUN}{function to apply}
}
\value{
A new `Features` object.

Calculate rolling features from forecasting package.

Calculate rolling features from forecasting package.
}
\description{
Calculate features specified in config file.

Create a new Features object.

Function calculates forecastas based on auto.arima and nnetar functions from forecast package.

Function calculates forecastas based on auto.arima and nnetar functions from forecast package.
https://stackoverflow.com/questions/38950005/how-to-manipulate-null-elements-in-a-nested-list/
}
