% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesDaily.R
\name{FeaturesDaily}
\alias{FeaturesDaily}
\title{Daily features}
\arguments{
\item{path}{Path to the data}

\item{features_config}{AWS S3 Tiledb config}

\item{at}{Row indecies for wich we calcilate the features. If NULL,
it will be calculated for all rows.}

\item{data}{X field of Ohlcv object}

\item{predictors_name}{Name of predictors}

\item{x}{List}

\item{FUN}{function to apply}
}
\value{
A new `FeaturesDaily` object.

Calculate rolling features from forecasting package.

Latest predictors

Calculate rolling features from forecasting package.
}
\description{
Calculate features specified in config file.

Create a new Features object.

Function calculates forecastas based on auto.arima and nnetar functions from forecast package.

Function get_latest get latest predictors from path.

Function calculates forecastas based on auto.arima and nnetar functions from forecast package.
https://stackoverflow.com/questions/38950005/how-to-manipulate-null-elements-in-a-nested-list/
}
