% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{finfeatures-package}
\alias{finfeatures}
\alias{finfeatures-package}
\title{finfeatures: Features for investing and finance}
\description{
Create and filter features that are helpfull in predictio problems in finance. Some basic features filtering techniques are also availabe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MislavSag/finfeatures}
  \item Report bugs at \url{https://github.com/MislavSag/finfeatures/issues}
}

}
