% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_finfeatures}
\alias{init_finfeatures}
\title{Communicate to R the Python virtual environment containing the relevant libraries for calculating features}
\usage{
init_finfeatures(
  python_path = "C:/Users/Mislav/.conda/envs/mlfinlabenv/python.exe"
)
}
\arguments{
\item{python_path}{\code{string} specifying the filepath to the Python virtual environment where "tsfresh", "tsfel", and/or "kats" are installed}
}
\value{
no return value; called for side effects
}
\description{
Communicate to R the Python virtual environment containing the relevant libraries for calculating features
}
\author{
Mislav Sagovac
}
