% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ohlcv.R
\name{Ohlcv}
\alias{Ohlcv}
\title{Ohlcv Class}
\description{
This function helps preparing data for feature calculation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X}}{Input data.}

\item{\code{id_col}}{I column.}

\item{\code{date_col}}{Date columnh}

\item{\code{price}}{Price column}

\item{\code{ohlcv}}{Ohlcv column}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ohlcv$new()}}
\item \href{#method-clone}{\code{Ohlcv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Ohlcv object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ohlcv$new(
  X,
  id_col = "symbol",
  date_col = "date",
  price = "close",
  ohlcv = c("open", "high", "low", "close", "volume")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Input data.}

\item{\code{id_col}}{Id column.}

\item{\code{date_col}}{Date column.}

\item{\code{price}}{Price column}

\item{\code{ohlcv}}{Ohlcv columns}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Person} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ohlcv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
