% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhlcvFeaturesDaily.R
\name{OhlcvFeaturesDaily}
\alias{OhlcvFeaturesDaily}
\title{OhlcvFeatures Class}
\arguments{
\item{at}{Row index for which we calcualte indicators.}

\item{windows}{Length of window for calculating rolling versions of the indicators.}

\item{quantile_divergence_window}{Window sizes from divergence from quantiles indicator.}

\item{ohlcv}{X field of Ohlcv object.}
}
\value{
A new `RollingGpd` object.

Data.table with new features.
}
\description{
Function calculates basic features from OHLCV financial data

Create a new OhlcvFeatures object.

Function calculates basic features from OHLCV financial data
}
\section{Fields}{

\describe{
\item{\code{Row}}{index for which we calcualte indicators.}

\item{\code{windows}}{Length of window for calculating rolling versions of the indicators.}

\item{\code{quantile_divergence_window}}{Window sizes from divergence from quantiles indicator.}
}}

