% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingBidAsk.R
\name{RollingBidAsk}
\alias{RollingBidAsk}
\title{RollingBidAsk Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{methods}{Argument methods in Bidask package.}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Prcie column in Ohlcv}
}
\value{
A new `RollingBidAsk` object.

Calculate rolling features from Bidask package.
}
\description{
Function calculates spread from exuber bidask package.

Create a new RollingBidAsk object.

Function calculates spread from exuber bidask package.
}
\section{Fields}{

\describe{
\item{\code{methods}}{Bidask methods.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingBidAskInstance <- RollingBidAsk$new(windows = 200,
                                           workers = 2L,
                                           at = c(300, 500),
                                           lag = 1L,
                                           na_pad = TRUE,
                                           simplify = FALSE
)
x = RollingBidAskInstance$get_rolling_features(OhlcvInstance)
head(x)
}
