% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RolingBinomialTrend.R
\name{RollingBinomialTrend}
\alias{RollingBinomialTrend}
\title{RollingBinomialTrend Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of threads.}

\item{lag}{Argument lag in runner package.}

\item{at}{Argument at in runner package.}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{price_col}{Prcie column in Ohlcv}

\item{params}{Vector of parameters}
}
\value{
A new `RollingBinomialTrend` object.

Calculate rolling radf features from exuber package.
}
\description{
Function calculates trend and p-value from  binomialtrend package

Create a new RollingBinomialTrend object.

Function calculates radf values from exuber package on rolling window.
}
\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingBinomialTrendInit = RollingBinomialTrend$new(windows = 200,
                                                    workers = 1L,
                                                    at = c(300, 500),
                                                    lag = 0L)
x = RollingBinomialTrendInit$get_rolling_features(OhlcvInstance)
head(x)
}
