% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingErf.R
\name{RollingErf}
\alias{RollingErf}
\title{Erf Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of threads.}

\item{lag}{Argument lag in runner package.}

\item{at}{Argument at in runner package.}

\item{m}{Argument m in vse4ts package.}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{price_col}{Prcie column in Ohlcv}

\item{params}{Vector of parameters}
}
\value{
A new `RollingVse` object.

Calculate rolling radf features from exuber package.
}
\description{
Function calculates predictions (quantiles) using extreme radnom forest
(erf package).

Create a new RollingErf object.

Function calculates radf values from exuber package on rolling window.
}
\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingErfInit = RollingErf$new(
  windows = 200,
  workers = 1L,
  at = c(300, 500),
  lag = 0L)
x = RollingErfInit$get_rolling_features(OhlcvInstance)
head(x)
}
