% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingFracdiff.R
\name{RollingFracdiff}
\alias{RollingFracdiff}
\title{RollingFracdiff Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of threads.}

\item{lag}{Argument lag in runner package.}

\item{at}{Argument at in runner package.}

\item{nar}{Argument nar in fracdiff function.}

\item{nma}{Argument nam in fracdiff function.}

\item{bandw_exp}{Argument bandw.exp in fracdiff function.}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{price_col}{Prcie column in Ohlcv}

\item{params}{Vector of parameters}
}
\value{
A new `RollingFracdiff` object.

Calculate rolling radf features from exuber package.
}
\description{
Function calculates coefficients of ARFIMA modelfrom fracdiff package.

Create a new RollingFracdiff object.

Function calculates radf values from exuber package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{nar}}{Argument nar in fracdiff function.}

\item{\code{nam}}{Argument nam in fracdiff function.}

\item{\code{bandw_exp}}{Argument bandw.exp in fracdiff function.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingFracdiffInit = RollingFracdiff$new(windows = 200,
                                          workers = 1L,
                                          at = c(300, 500),
                                          lag = 0L,
                                          nar = 1,
                                          nma = 1,
                                          bandw_exp = c(0.1, 0.2))
x = RollingFracdiffInit$get_rolling_features(OhlcvInstance)
head(x)
}
