% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingGpd.R
\name{RollingGpd}
\alias{RollingGpd}
\title{RollingGpd Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{threshold}{Threshold parameter for returns. Returns below/above this threshold will be input.}

\item{hill_threshold}{look at ptsuite package}

\item{suffix}{suffix for column names}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{price_col}{Prcie column in Ohlcv}

\item{params}{Vector of parameters}
}
\value{
A new `RollingGpd` object.

Data.table with new features

Calculate rolling radf features from exuber package.
}
\description{
Function calculates GPD risk measures on rolling window from ptsuite pakcage.

Create a new RollingGpd object.

Help function for calculating GPD features.

Function calculates radf values from exuber package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{threshold}}{Threshold parameter for returns. Returns below/above this threshold will be input.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingGpdInit = RollingGpd$new(windows = 200,
                                workers = 1L,
                                at = c(300, 500),
                                lag = 0L,
                                threshold = 0.03)
x = RollingGpdInit$get_rolling_features(OhlcvInstance)
head(x)
RollingGpdInit = RollingGpd$new(windows = 200,
                                workers = 2L,
                                at = c(300, 500),
                                lag = 0L,
                                threshold = c(0,02, 0.03))
x = RollingGpdInit$get_rolling_features(OhlcvInstance)
head(x)
}
