% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingWaveletArima.R
\name{RollingWaveletArima}
\alias{RollingWaveletArima}
\title{RollingWaveletArima Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of threads.}

\item{lag}{Argument lag in runner package.}

\item{at}{Argument at in runner package.}

\item{filter}{Wavelet filter use in the decomposition.}

\item{Waveletlevels}{The level of wavelet decomposition.}

\item{MaxARParam}{The maximum AR order for auto.arima.}

\item{MaxMAParam}{The maximum MA order for auto.arima.}

\item{NForecast}{Forecast horizont.}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{price_col}{Prcie column in Ohlcv}

\item{params}{Vector of parameters}
}
\value{
A new `RollingForecats` object.

Calculate rolling radf features from exuber package.
}
\description{
Function calculates forecastas based on WaveletArima package.

Create a new RollingForecats object.

Function calculates radf values from exuber package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{filter}}{Wavelet filter use in the decomposition.}

\item{\code{Waveletlevels}}{The level of wavelet decomposition.}

\item{\code{MaxARParam}}{The maximum AR order for auto.arima.}

\item{\code{MaxMAParam}}{The maximum MA order for auto.arima.}

\item{\code{NForecast}}{Forecast horizont.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingWaveletArimaInit = RollingWaveletArima$new(windows = 200,
                                                  workers = 1L,
                                                  at = c(300, 500),
                                                  lag = 0L,
                                                  filter = "haar")
x = RollingWaveletArimaInit$get_rolling_features(OhlcvInstance)
head(x)
RollingWaveletArimaInit = RollingWaveletArima$new(windows = 200,
                                                  workers = 2L,
                                                  at = c(300, 500),
                                                  lag = 0L,
                                                  filter = c("haar", "la8"))
x = RollingWaveletArimaInit$get_rolling_features(OhlcvInstance)
head(x)
}
