% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy_hour.R
\docType{data}
\name{spy_hour}
\alias{spy_hour}
\title{SPY market data}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{datetime}{Date and time}
 \item{open}{Open price}
 \item{high}{High price}
 \item{low}{Low price}
 \item{close}{Close price}
 \item{volume}{Volumme}
}
}
\usage{
data(spy_hour)
}
\description{
Market data for SPY with hour frequency for period from 2004-01-01 to 2021-10-20.
Data was generated through FMP cloud Prep API.
}
\examples{

data(spy_hour)
head(spy_hour)
}
\references{
Data was collected using FMP Cloud Prep API
}
\keyword{datasets}
