% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Stocks market data}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{datetime}{Date and time}
 \item{open}{Open price}
 \item{high}{High price}
 \item{low}{Low price}
 \item{close}{Close price}
 \item{volume}{Volumme}
}
}
\usage{
data(stocks)
}
\description{
Market data for sample of stocks with hour frequency for period from
2004-01-01 to 2021-10-20. Data was generated through FMP cloud Prep API.
}
\examples{

data(stocks)
head(stocks)
}
\references{
Data was collected using FMP Cloud Prep API
}
\keyword{datasets}
