% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{Features}
\alias{Features}
\title{Features Class}
\arguments{
\item{ohlcv_object}{Length of window for calculating rolling versions of the indicators.}

\item{at}{Argument `at` in runner package}

\item{lag}{Argument `lag` in runner package}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}
}
\value{
A new `Features` object.

Data.table with new features.
}
\description{
Calculates all features for financial time series forecasting.

Create a new OhlcvFeatures object.

Function calculates basic features from OHLCV financial data
}
\section{Fields}{

\describe{
\item{\code{ohlcv_object}}{OHLCV object}

\item{\code{at}}{Argument `at` in runner package}

\item{\code{lag}}{Argument `lag` in runner package}
}}

