% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ohlcv.R
\name{Ohlcv}
\alias{Ohlcv}
\title{Ohlcv Class}
\arguments{
\item{X}{Input data.}

\item{id_col}{Id column.}

\item{date_col}{Date column.}

\item{price}{Price column}

\item{ohlcv}{Ohlcv columns}
}
\value{
A new `Person` object.
}
\description{
This function helps preparing data for feature calculation.

Create a new Ohlcv object.
}
\section{Fields}{

\describe{
\item{\code{X}}{Input data.}

\item{\code{id_col}}{I column.}

\item{\code{date_col}}{Date columnh}

\item{\code{price}}{Price column}

\item{\code{ohlcv}}{Ohlcv column}
}}

