% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhlcvFeatures.R
\name{OhlcvFeatures}
\alias{OhlcvFeatures}
\title{OhlcvFeatures Class}
\arguments{
\item{windows}{Length of window for calculating rolling versions of the indicators.}

\item{quantile_divergence_window}{Window sizes from divergence from quantiles indicator.}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{returns}{Returns column.}

\item{method}{Argument method from cpm package}

\item{ARL0}{Argument ARL0 from cpm package}

\item{startup}{Argument sartup from cpm package}
}
\value{
A new `RollingGpd` object.

Data.table with new features.

Data.table with new features.
}
\description{
Function calculates basic features from OHLCV financial data

Create a new OhlcvFeatures object.

Function calculates basic features from OHLCV financial data

Helper function for calculating changepoints features
}
\section{Fields}{

\describe{
\item{\code{windows}}{Length of window for calculating rolling versions of the indicators.}

\item{\code{quantile_divergence_window}}{Window sizes from divergence from quantiles indicator.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingGpdInit = OhlcvFeatures$new(windows = c(200, 300),
                                   quantile_divergence_window =  c(50, 100))
x = RollingGpdInit$get_ohlcv_features(OhlcvInstance)
tail(x)
}
