% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingBackcusum.R
\name{RollingBackcusum}
\alias{RollingBackcusum}
\title{RollingBackcusum Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of threads.}

\item{lag}{Argument lag in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Prcie column in Ohlcv}
}
\value{
A new `RollingBackcusum` object.

Calculate rolling radf features from backCUSUM package.
}
\description{
Function calculates backCUSUM tests on rolling window from backCUSUM pakcage.

Create a new RollingBackcusum object.

Function calculates radf values from exuber package on rolling window.
}
\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingBackcusumInit = RollingBackcusum$new(windows = 200,
                                            workers = 1L,
                                            at = c(300, 500),
                                            lag = 0L,
                                            na_pad = TRUE,
                                            simplify = FALSE)
x = RollingBackcusumInit$get_rolling_features(OhlcvInstance)
head(x)
# multiple windows and parallel
}
