% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RolingBinomialTrend.R
\name{RollingBinomialTrend}
\alias{RollingBinomialTrend}
\title{RollingBinomialTrend Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of threads.}

\item{lag}{Argument lag in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Price column in Ohlcv}
}
\value{
A new `RollingBinomialTrend` object.

Calculate rolling trend, p-value features from binomialtrend package.
}
\description{
Function calculates trend and p-value from  binomialtrend package

Create a new RollingBinomialTrend object.

Function calculates trend, p-value features from binomialtrend on rolling window.
}
\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingBinomialTrendInit = RollingBinomialTrend$new(windows = 200,
                                            workers = 1L,
                                            at = c(300, 500),
                                            lag = 0L,
                                            na_pad = TRUE,
                                            simplify = FALSE)
x = RollingBinomialTrendInit$get_rolling_features(OhlcvInstance)
head(x)
# multiple windows and parallel
}
