% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingFracdiff.R
\name{RollingFracdiff}
\alias{RollingFracdiff}
\title{RollingFracdiff Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of threads.}

\item{lag}{Argument lag in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{nar}{Argument nar in fracdiff function.}

\item{nma}{Argument nam in fracdiff function.}

\item{bandw_exp}{Argument bandw.exp in fracdiff function.}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Prcie column in Ohlcv}
}
\value{
A new `RollingFracdiff` object.

Calculate rolling features from forecasting package.
}
\description{
Function calculates coefficients of ARFIMA modelfrom fracdiff package.

Create a new RollingFracdiff object.

Function calculates forecastas based on auto.arima and nnetar functions from forecast package.
}
\section{Fields}{

\describe{
\item{\code{nar}}{Argument nar in fracdiff function.}

\item{\code{nam}}{Argument nam in fracdiff function.}

\item{\code{bandw_exp}}{Argument bandw.exp in fracdiff function.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
#arima
RollingForecatsInstance = RollingForecats$new(windows = c(10, 20),
                                              workers = 1L,
                                              lag = 1L,
                                              at = c(100:110, 200:210),
                                              na_pad = TRUE,
                                              simplify = FALSE,
                                              forecast_type = "autoarima",
                                              h = 22)
x = RollingForecatsInstance$get_rolling_features(OhlcvInstance)
head(x)
}
