% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingGas.R
\name{RollingGas}
\alias{RollingGas}
\title{RollingGas Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{gas_dist}{Dist parametere in UniGASSpec fucntionUniGASSpec.}

\item{gas_scaling}{Scaling parametere in UniGASSpec fucntionUniGASSpec.}

\item{prediction_horizont}{GAS prediction horizont.}

\item{df}{A data.table object with GAS results.}

\item{colname_prefix}{Prefix for column names}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Prcie column in Ohlcv}
}
\value{
A new `RollingGas` object.

Calculate rolling GAS features from GAS package.

Calculate rolling GAS features from GAS package.
}
\description{
Function calculates GAS risk values from GAS package on rolling window.

Create a new RollingExuber object.

Help function for Rolling GAS

Function calculates GAS risk values from GAS package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{gas_dist}}{Dist parametere in UniGASSpec fucntionUniGASSpec.}

\item{\code{gas_scaling}}{Scaling parametere in UniGASSpec fucntionUniGASSpec.}

\item{\code{prediction_horizont}}{GAS prediction horizont.}

\item{\code{GASSpec}}{Defined internally}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingGasInit = RollingGas$new(windows = 200,
                                workers = 1L,
                                at = c(300, 500),
                                lag = 0L,
                                na_pad = TRUE,
                                simplify = FALSE,
                                gas_dist = "sstd",
                                gas_scaling = "Identity",
                                prediction_horizont = 10)
x = RollingGasInit$get_rolling_features(OhlcvInstance)
head(x)
# parallel and multiple windows
RollingGasInit = RollingGas$new(windows = c(200, 300),
                                workers = 4L,
                                at = c(300:310, 500:510),
                                lag = 1L,
                                na_pad = TRUE,
                                simplify = FALSE,
                                gas_dist = "sstd",
                                gas_scaling = "Identity",
                                prediction_horizont = 10)
x = RollingGasInit$get_rolling_features(OhlcvInstance)
head(x)
}
