% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingGeneric.R
\name{RollingGeneric}
\alias{RollingGeneric}
\title{RollingGeneric Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{packages}{Packages that need to be installed.}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{params}{Vector of parameters}

\item{Ohlcv}{Ohlcv object}

\item{log_prices}{If TRUE, we log OHLC.}

\item{price_col}{Price column in Ohlcv}
}
\value{
A new `RollingGeneric` object.

Depending on used classes for clalculating features. It returns df.

Depending on used classes for clalculating features. It returns df.
}
\description{
This is the abstract base class for tasks that start with "Rolling".
It helps calculating rolling values of features.

Create a new RollingGeneric object.

This function will be used in application. For now it doesn't do anythin

Helping function for calculating rolling features.
}
\section{Fields}{

\describe{
\item{\code{windows}}{Vector of windows that will be applied on features.}

\item{\code{workers}}{Number of workers. Greater than 1 for parallle processing}

\item{\code{lag}}{Lag variable in runner package.}

\item{\code{at}}{Argument at in runner package.}

\item{\code{packages}}{Packages that needs to be installed.}
}}

