% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingGpd.R
\name{RollingGpd}
\alias{RollingGpd}
\title{RollingGpd Class}
\description{
Function calculates GPD risk measures on rolling window from ptsuite pakcage.
}
\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingGpdInit = RollingGpd$new(windows = 200,
                                workers = 1L,
                                at = c(300, 500),
                                lag = 0L,
                                na_pad = TRUE,
                                simplify = FALSE,
                                threshold = 0.01)
x = RollingGpdInit$get_rolling_features(OhlcvInstance)
head(x)
# OHLCV features
# parallel and multiple windows
RollingGasInit = RollingGpd$new(windows = c(200, 300),
                                workers = 4L,
                                at = c(300:310, 500:510),
                                lag = 1L,
                                na_pad = TRUE,
                                simplify = FALSE,
                                threshold = 0.01)
x = RollingGasInit$get_rolling_features(OhlcvInstance)
head(x)
}
\section{Super class}{
\code{\link[finfeatures:RollingGeneric]{finfeatures::RollingGeneric}} -> \code{RollingGpd}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{threshold}}{Threshold parameter for returns. Returns below/above this threshold will be input.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RollingGpd$new()}}
\item \href{#method-estimate_gpd}{\code{RollingGpd$estimate_gpd()}}
\item \href{#method-rolling_function}{\code{RollingGpd$rolling_function()}}
\item \href{#method-clone}{\code{RollingGpd$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="finfeatures" data-topic="RollingGeneric" data-id="get_rolling_features">}\href{../../finfeatures/html/RollingGeneric.html#method-get_rolling_features}{\code{finfeatures::RollingGeneric$get_rolling_features()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new RollingGpd object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollingGpd$new(windows, workers, lag, at, na_pad, simplify, threshold = 0.03)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{windows}}{Vector of windows that will be applied on features.}

\item{\code{workers}}{Number of workers. Greater than 1 for parallle processing}

\item{\code{lag}}{Lag variable in runner package.}

\item{\code{at}}{Argument at in runner package.}

\item{\code{na_pad}}{Argument na_pad in runner package.}

\item{\code{simplify}}{Argument simplify in runner package.}

\item{\code{threshold}}{Threshold parameter for returns. Returns below/above this threshold will be input.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{RollingGpd} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate_gpd"></a>}}
\if{latex}{\out{\hypertarget{method-estimate_gpd}{}}}
\subsection{Method \code{estimate_gpd()}}{
Help function for calculating GPD features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollingGpd$estimate_gpd(x, hill_threshold = 0.02, suffix = "_right")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of (filtered) returns}

\item{\code{hill_threshold}}{look at ptsuite package}

\item{\code{suffix}}{suffix for column names}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data.table with new features
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rolling_function"></a>}}
\if{latex}{\out{\hypertarget{method-rolling_function}{}}}
\subsection{Method \code{rolling_function()}}{
Function calculates GPD risk measures on rolling window from ptsuite pakcage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollingGpd$rolling_function(data, window)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{X field of Ohlcv object}

\item{\code{window}}{window length. This argument is given internaly}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Calculate rolling features from ptsuite package.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RollingGpd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
