% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingGpd.R
\name{RollingGpd}
\alias{RollingGpd}
\title{RollingGpd Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{threshold}{Threshold parameter for returns. Returns below/above this threshold will be input.}

\item{x}{vector of (filtered) returns}

\item{hill_threshold}{look at ptsuite package}

\item{suffix}{suffix for column names}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Prcie column in Ohlcv}
}
\value{
A new `RollingGpd` object.

Data.table with new features

Calculate rolling features from ptsuite package.
}
\description{
Function calculates GPD risk measures on rolling window from ptsuite pakcage.

Create a new RollingGpd object.

Help function for calculating GPD features.

Function calculates GPD risk measures on rolling window from ptsuite pakcage.
}
\section{Fields}{

\describe{
\item{\code{threshold}}{Threshold parameter for returns. Returns below/above this threshold will be input.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingGpdInit = RollingGpd$new(windows = 200,
                                workers = 1L,
                                at = c(300, 500),
                                lag = 0L,
                                na_pad = TRUE,
                                simplify = FALSE,
                                threshold = 0.01)
x = RollingGpdInit$get_rolling_features(OhlcvInstance)
head(x)
RollingGpdInit = RollingGpd$new(windows = 200,
                                workers = 2L,
                                at = c(300:310, 500:510),
                                lag = 0L,
                                na_pad = TRUE,
                                simplify = FALSE,
                                threshold = 0.01)
x = RollingGpdInit$get_rolling_features(OhlcvInstance)
head(x)
}
