% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingQuarks.R
\name{RollingQuarks}
\alias{RollingQuarks}
\title{RollingQuarks Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{p}{Argument p in quarks package}

\item{model}{Argument model in quarks package}

\item{method}{Argument method in quarks package}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Prcie column in Ohlcv}
}
\value{
A new `RollingQuarks` object.

Calculate rolling radf features from exuber package.
}
\description{
Function calculates risk measures (Var, ES) from quarks package on rolling window.

Create a new RollingExuber object.

Function calculates risk measures (Var, ES) from quarks package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{p}}{Argument p in quarks package}

\item{\code{model}}{Argument model in quarks package}

\item{\code{method}}{Argument method in quarks package}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingQuarksInit = RollingQuarks$new(windows = 200,
                                      workers = 1L,
                                      at = c(300:310, 500:510),
                                      lag = 1L,
                                      na_pad = TRUE,
                                      simplify = FALSE)
x = RollingQuarksInit$get_rolling_features(OhlcvInstance)
head(x)
# parallel and multiple windows
RollingQuarksInit = RollingQuarks$new(windows = c(200, 400),
                                      workers = 2L,
                                      at = c(300:310, 500:510),
                                      lag = 1L,
                                      na_pad = TRUE,
                                      simplify = FALSE)
x = RollingQuarksInit$get_rolling_features(OhlcvInstance)
head(x)
}
