% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingTsfeatures.R
\name{RollingTsfeatures}
\alias{RollingTsfeatures}
\title{RollingTsfeatures Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{scale}{Check argument of tsfeatures function.}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{price_col}{Prcie column in Ohlcv}

\item{params}{Vector of parameters}
}
\value{
A new `RollingTsfeatures` object.

Calculate rolling radf features from exuber package.
}
\description{
Function calculates tsfeatures features on rolling window.

Create a new RollingTsfeatures object.

Function calculates radf values from exuber package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{scale}}{Check argument of tsfeatures function.}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingTsfeaturesInit = RollingTsfeatures$new(windows = 200,
                                              workers = 1L,
                                              at = c(300, 500),
                                              lag = 0L,
                                              scale = TRUE)
x = RollingTsfeaturesInit$get_rolling_features(OhlcvInstance)
head(x)
RollingTsfeaturesInit = RollingTsfeatures$new(windows = 200,
                                              workers = 2L,
                                              at = c(300, 500),
                                              lag = 0L,
                                              scale = c(TRUE, FALSE))
x = RollingTsfeaturesInit$get_rolling_features(OhlcvInstance)
head(x)
}
