% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingTvgarch.R
\name{RollingTvgarch}
\alias{RollingTvgarch}
\title{RollingTvgarch Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{turbo}{Check argument in tvgarch function.}

\item{x}{Ohlcv object.}

\item{window}{Rolling window lengths.}

\item{price_col}{Prcie column in Ohlcv}

\item{params}{Vector of parameters}
}
\value{
A new `RollingGas` object.

Calculate rolling radf features from exuber package.
}
\description{
Function calculates tvgarch predictions and coefficients from tvgarch package on rolling window.

Create a new RollingExuber object.

Function calculates radf values from exuber package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{turbo}}{Check argument in tvgarch function.}
}}

