% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingUfrisk.R
\name{RollingUfrisk}
\alias{RollingUfrisk}
\title{RollingUfrisk Class}
\arguments{
\item{windows}{Vector of windows that will be applied on features.}

\item{workers}{Number of workers. Greater than 1 for parallle processing}

\item{lag}{Lag variable in runner package.}

\item{at}{Argument at in runner package.}

\item{na_pad}{Argument na_pad in runner package.}

\item{simplify}{Argument simplify in runner package.}

\item{a_v}{Argument a.v in ufRisk package}

\item{a_e}{Argument a.e in ufRisk package}

\item{model}{Argument model in ufRisk package}

\item{garch_order}{Argument garch_order in ufRisk package}

\item{data}{X field of Ohlcv object}

\item{window}{window length. This argument is given internaly}

\item{price}{Prcie column in Ohlcv}
}
\value{
A new `RollingQuarks` object.

Calculate rolling Var and ES features from ufRisk package.
}
\description{
Function calculates risk measures (Var, ES) from ufRisk package on rolling window.

Create a new RollingExuber object.

Function calculates risk measures (Var, ES) from ufRisk package on rolling window.
}
\section{Fields}{

\describe{
\item{\code{a_v}}{Argument a.v in ufRisk package}

\item{\code{a_e}}{Argument a.e in ufRisk package}

\item{\code{model}}{Argument model in ufRisk package}

\item{\code{garch_order}}{Argument garchOrder in ufRisk package}
}}

\examples{
data(spy_hour)
OhlcvInstance = Ohlcv$new(spy_hour, date_col = "datetime")
RollingUfriskInit = RollingUfrisk$new(windows = 200,
                                      workers = 1L,
                                      at = c(300:310, 500:510),
                                      lag = 1L,
                                      na_pad = TRUE,
                                      simplify = FALSE)
x = RollingUfriskInit$get_rolling_features(OhlcvInstance)
head(x)
# parallel and multiple windows
RollingUfriskInit = RollingUfrisk$new(windows = c(200, 400),
                                      workers = 2L,
                                      at = c(300:310, 500:510),
                                      lag = 1L,
                                      na_pad = TRUE,
                                      simplify = FALSE)
x = RollingUfriskInit$get_rolling_features(OhlcvInstance)
head(x)
}
